Add more ergonomic ways to deal with uefi-rs' Results
